+++ src/app/(route)/(withLayout)/mypage/layout.tsx
@@ -1,17 +1,19 @@
+import RQProvider from '@/app/_components/RQProvider'
+
 export default function MypageLayout({
   children,
 }: {
   children: React.ReactNode
 }) {
   return (
-    <div
+    <main
       style={{
         backgroundImage:
           'linear-gradient(to bottom, white 107px, #F7F7F7 107px)',
       }}
       className="flex w-full px-[150px] py-[49px]"
     >
-      {children}
-    </div>
+      <RQProvider>{children}</RQProvider>
+    </main>
   )
 }
